CasePraticoAutoU/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.2",
    "openai>=2.1.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
